Next iteration of volume calculation
